+++ CHANGELOG.md
@@ -52,7 +52,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Remove deprecated `GraphQL::executeAndReturnResult()`
 - Remove deprecated experimental CoroutineExecutor
 - Remove deprecated `FormattedError::create()` and `FormattedError::createFromPHPError()`
-- Remove `GraphQL::setPromiseAdapter()`
+- Remove deprecated `GraphQL::setPromiseAdapter()`
 - Remove deprecated `AST::getOperation()`
 - Remove deprecated constants from `BreakingChangesFinder`
 - Remove deprecated `DocumentValidator::isValidLiteralValue()`
@@ -61,6 +61,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Remove deprecated `Schema::isPossibleType()`
 - Remove deprecated methods from `TypeInfo`
 - Remove deprecated `Values::valueFromAST()` and `Values::isValidPHPValue()`
+- Remove deprecated public property access to `InputObjectField::$type`
 
 ## 14.9.0